# dormitory-backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dormitory API",
    "description": "API for managing dormitory operations including authentication, admin, areas, registrations, reports, rooms, users, contracts, bills, notifications, services, and payments",
    "version": "1.0.0"
  },
  "paths": {
...
  "/api/service-rates": {
    "get": {
      "tags": ["ServiceRate"],
      "summary": "Get all service rates (Admin)",
      "description": "Retrieve a paginated list of service rates, optionally filtered by service ID. Requires admin privileges.",
      "security": [{"Bearer": []}],
      "parameters": [
        {
          "name": "page",
          "in": "query",
          "schema": {"type": "integer", "default": 1},
          "description": "Page number for pagination"
        },
        {
          "name": "limit",
          "in": "query",
          "schema": {"type": "integer", "default": 10},
          "description": "Number of items per page"
        },
        {
          "name": "service_id",
          "in": "query",
          "schema": {"type": "integer"},
          "description": "Filter by service ID"
        }
      ],
      "responses": {
        "200": {
          "description": "List of service rates",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "service_rates": {
                    "type": "array",
                    "items": {"$ref": "#/components/schemas/ServiceRate"}
                  },
                  "total": {"type": "integer", "example": 100},
                  "pages": {"type": "integer", "example": 10},
                  "current_page": {"type": "integer", "example": 1}
                }
              }
            }
          }
        },
        "404": {
          "description": "Service not found",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Không tìm thấy dịch vụ với ID 999"
                  }
                }
              }
            }
          }
        },
        "403": {
          "description": "Admin access required",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Yêu cầu quyền admin"
                  }
                }
              }
            }
          }
        },
        "500": {
          "description": "Server error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Lỗi khi lấy danh sách mức giá"
                  },
                  "error": {"type": "string"}
                }
              }
            }
          }
        }
      }
    },
    "post": {
      "tags": ["ServiceRate"],
      "summary": "Create a new service rate (Admin)",
      "description": "Create a new service rate for a service. Requires admin privileges.",
      "security": [{"Bearer": []}],
      "requestBody": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "service_id": {
                  "type": "integer",
                  "example": 1,
                  "description": "ID of the service"
                },
                "unit_price": {
                  "type": "string",
                  "format": "decimal",
                  "example": "5000.00",
                  "description": "Price per unit, non-negative, max 99999999.99"
                },
                "effective_date": {
                  "type": "string",
                  "format": "date",
                  "example": "2025-04-01",
                  "description": "Effective date in YYYY-MM-DD format"
                }
              },
              "required": ["service_id", "unit_price", "effective_date"]
            }
          }
        }
      },
      "responses": {
        "201": {
          "description": "Service rate created",
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ServiceRate"}
            }
          }
        },
        "400": {
          "description": "Invalid input",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "enum": [
                      "Yêu cầu service_id, unit_price và effective_date",
                      "service_id phải là số nguyên",
                      "unit_price phải là số không âm và không vượt quá 99999999.99",
                      "Định dạng effective_date không hợp lệ (YYYY-MM-DD)",
                      "Dữ liệu JSON không hợp lệ"
                    ]
                  }
                }
              }
            }
          }
        },
        "404": {
          "description": "Service not found",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Không tìm thấy dịch vụ với ID 999"
                  }
                }
              }
            }
          }
        },
        "403": {
          "description": "Admin access required",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Yêu cầu quyền admin"
                  }
                }
              }
            }
          }
        },
        "500": {
          "description": "Server error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Lỗi khi tạo mức giá"
                  },
                  "error": {"type": "string"}
                }
              }
            }
          }
        }
      }
    }
  },
...
  "/api/service-rates/current/{service_id}": {
    "get": {
      "tags": ["ServiceRate"],
      "summary": "Get current service rate (Admin)",
      "description": "Retrieve the current service rate for a service based on effective date. Requires admin privileges.",
      "security": [{"Bearer": []}],
      "parameters": [
        {
          "name": "service_id",
          "in": "path",
          "required": true,
          "schema": {"type": "integer"},
          "description": "ID of the service"
        }
      ],
      "responses": {
        "200": {
          "description": "Current service rate",
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ServiceRate"}
            }
          }
        },
        "404": {
          "description": "Service or current rate not found",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "enum": [
                      "Không tìm thấy dịch vụ với ID {service_id}",
                      "Không tìm thấy mức giá hiện tại cho dịch vụ này"
                    ]
                  }
                }
              }
            }
          }
        },
        "403": {
          "description": "Admin access required",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Yêu cầu quyền admin"
                  }
                }
              }
            }
          }
        },
        "500": {
          "description": "Server error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Lỗi khi lấy mức giá hiện tại"
                  },
                  "error": {"type": "string"}
                }
              }
            }
          }
        }
      }
    }
  },
  "/api/service-rates/{rate_id}": {
    "get": {
      "tags": ["ServiceRate"],
      "summary": "Get service rate by ID (Admin)",
      "description": "Retrieve details of a service rate by ID. Requires admin privileges.",
      "security": [{"Bearer": []}],
      "parameters": [
        {
          "name": "rate_id",
          "in": "path",
          "required": true,
          "schema": {"type": "integer"},
          "description": "ID of the service rate"
        }
      ],
      "responses": {
        "200": {
          "description": "Service rate details",
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ServiceRate"}
            }
          }
        },
        "404": {
          "description": "Rate not found",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Không tìm thấy mức giá với ID 999"
                  }
                }
              }
            }
          }
        },
        "403": {
          "description": "Admin access required",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Yêu cầu quyền admin"
                  }
                }
              }
            }
          }
        },
        "500": {
          "description": "Server error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Lỗi khi lấy mức giá"
                  },
                  "error": {"type": "string"}
                }
              }
            }
          }
        }
      }
    },
    "put": {
      "tags": ["ServiceRate"],
      "summary": "Update service rate (Admin)",
      "description": "Update an existing service rate. Requires admin privileges.",
      "security": [{"Bearer": []}],
      "parameters": [
        {
          "name": "rate_id",
          "in": "path",
          "required": true,
          "schema": {"type": "integer"},
          "description": "ID of the service rate"
        }
      ],
      "requestBody": {
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "service_id": {
                  "type": "integer",
                  "example": 2,
                  "description": "ID of the service"
                },
                "unit_price": {
                  "type": "string",
                  "format": "decimal",
                  "example": "6000.00",
                  "description": "Price per unit, non-negative, max 99999999.99"
                },
                "effective_date": {
                  "type": "string",
                  "format": "date",
                  "example": "2025-05-01",
                  "description": "Effective date in YYYY-MM-DD format"
                }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Service rate updated",
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ServiceRate"}
            }
          }
        },
        "400": {
          "description": "Invalid input",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "enum": [
                      "service_id phải là số nguyên",
                      "unit_price phải là số không âm và không vượt quá 99999999.99",
                      "Định dạng effective_date không hợp lệ (YYYY-MM-DD)",
                      "Dữ liệu JSON không hợp lệ"
                    ]
                  }
                }
              }
            }
          }
        },
        "404": {
          "description": "Rate or service not found",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "enum": [
                      "Không tìm thấy mức giá với ID {rate_id}",
                      "Không tìm thấy dịch vụ với ID {service_id}"
                    ]
                  }
                }
              }
            }
          }
        },
        "403": {
          "description": "Admin access required",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Yêu cầu quyền admin"
                  }
                }
              }
            }
          }
        },
        "500": {
          "description": "Server error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Lỗi khi cập nhật mức giá"
                  },
                  "error": {"type": "string"}
                }
              }
            }
          }
        }
      }
    },
    "delete": {
      "tags": ["ServiceRate"],
      "summary": "Delete service rate (Admin)",
      "description": "Delete a service rate by ID. Requires admin privileges.",
      "security": [{"Bearer": []}],
      "parameters": [
        {
          "name": "rate_id",
          "in": "path",
          "required": true,
          "schema": {"type": "integer"},
          "description": "ID of the service rate"
        }
      ],
      "responses": {
        "204": {
          "description": "Service rate deleted"
        },
        "404": {
          "description": "Rate not found",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Không tìm thấy mức giá với ID 999"
                  }
                }
              }
            }
          }
        },
        "403": {
          "description": "Admin access required",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Yêu cầu quyền admin"
                  }
                }
              }
            }
          }
        },
        "500": {
          "description": "Server error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Lỗi khi xóa mức giá"
                  },
                  "error": {"type": "string"}
                  }
                }
             }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServiceRate": {
        "type": "object",
        "properties": {
          "rate_id": {"type": "integer"},
          "service_id": {"type": "integer"},
          "unit_price": {"type": "number"},
          "effective_date": {"type": "string", "format": "date"}
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}